Create SessionIOS factory which can be passed to SessionService
SessionService has a savesession method which used to take a code
block that creates a session and execute with a delay. The problem with
that is it depended on TabModel existence to check if it can create a
session for saving or not.

Instead create a real SessionIOSFactory for now it's owned by TabModel,
but later it will be owned by SessionRestorationAgent. The SessionIOSFactory
will have a webStateList and will create a sessionIOS object serializing this
webStateList when it's asked for. It's assuming that as long as the object for the
SessionIOSFactory is alive, then the member webStateList will be alive. So it can
use it directly to create a session whenever it needs to.
SessionService now takes the factory as a weak pointer and it doesn't retain it.
So if the owner is deallocated  (now it's the TabModel, and it will be
SessionRestorationAgent later) then the factory will be nil and it will not create
a session.

Bug: 1010164

Change-Id: Ie168fc26a57997be59e123f4ac449d063a3de929
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1953084
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Mohammad Refaat <mrefaat@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#723364}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fa8014ef3e77ad564c3cb9f07c74f861e1e2f4d8