[ios] Fix language selection bug in Translate Messages
This change adds the other language when listing languages to change
either the source or target language. This ensures that the row number
of the language matches its index in the TranslateInfobarDelegate's
model. Otherwise, the language associated with the index would not
properly match the language selected.

Bug: 1014959
Change-Id: I7703bac3ff959e52ef715cfc1a135faffa3582ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1959471
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Commit-Queue: Chris Lu <thegreenfrog@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#724432}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 348019a04a49d69e6a29229fc30f2660eecdb9f8