[iOS] Create InfobarOverlayType enum.
InfobarOverlayType is used to describe the types of OverlayRequest
configurations that will be supported for a single InfoBar.  These are
used to show overlays for the banner UI (kBanner), the action sheet
displayed in Messages 2.0 showing detailed options (kDetailSheet), and
the modal UI used to update settings for that infobar (kModal).

InfobarOverlayRequestFactory was updated to take an InfobarOverlayType
in its factory method to allow the creation of OverlayRequests
configured for each of these overlay types for a given InfoBar.
InfobarOverlayRequestFactoryImpl was updated to use templated helpers
to create requests instead of relying on switch statements and boolean
logic to decide which configuration to use for each InfobarType
and InfobarOverlayType pair.

Bug: 1030357
Change-Id: Iaba54c95dfece2381164ccf8a615b8fda28c42d7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1959936
Commit-Queue: Kurt Horimoto <kkhorimoto@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#724796}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 237e4fad6e69aee6c985e9df01c74a988b1b469c