[iOS] Create TabInsertionBrowserAgent
This CL factors out the tab insertion logic from TabModel into a
Browser agent. In order to do this, it also:

 - Creates the BrowserUserData infrastructure.
 - Creates an AttachBrowserAgents() helper.

These are direct analogs to WebStateUserData and AttachTabHelpers().

These TabModel methods are removed:

-insertWebStateWithURL:referrer:transition:opener:openedByDOM:atIndex:inBackground:
-insertWebStateWithLoadParams:opener:openedByDOM:atIndex:inBackground:
-insertOpenByDOMWebStateWithOpener:

The first of these was a convenience wrapper for the second (it used the
first three parameters to construct the LoadParams: argument for the wrapped
method). It was mostly used in tests (which now use a separate helper), and the
one non-test callsite (in BVC) just creates the LoadParams explicitly.

The other two methods are replaced by:
  TabInsertionBrowserAgent::InsertWebState
  TabInsertionBrowserAgent::InsertWebStateOpenedByDOM

Callsites generally grab the TabInsertionBrowserAgent from a nearby Browser. Only
TabGridAdaptor needed to have plumbing added for this (in the TabSwitcher protocol).

The bulk of the changed lines come from (a) moving more than half of TabModel's
unit tests into TabInsertionBrowserAgentTest, and (b) updating the ObjC method
calls (again, mostly in tests) to use the new C++ calls.

Change-Id: I1aaeebab0cba49801269d6531faba16086563f93
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1969890
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Commit-Queue: Mark Cogan <marq@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#725545}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: dc90dd1f05642eed5614516d0dc8d045105fa4f6