[iOS] Reenable testPasswordIconIsNotVisibleWhenPasswordStoreEmpty
On iOS 13.2 this test started to fail on EG1. EG evaluates the icon as not
visible, even thought it is clearly in the view hierarchy. I wasn't able to
fix it after spending some hours on it, removing the second part of the
test, since the first part already test the purpose of it.

Bug: 1019535
Change-Id: If7da63916e9fe9ea757a129358d101372e3fa032
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1975159
Reviewed-by: Chris Lu <thegreenfrog@chromium.org>
Commit-Queue: Javier Ernesto Flores Robles <javierrobles@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#727177}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ff0e0db67b0e777ed0076c5e532806cd327aaa86