[iOS] Resize the NTP instead of adjusting ContentInset
This CL resizes the frame of the NTP instead of adjusting its content
inset. It allows the content of the NTP to not be displayed below the
toolbar, avoid issues with VoiceOver reading the item behind it.
It also allows to fix the content insets for the safe area.

Fixed: 1034509, 1030422
Change-Id: I306ecf5b771fa3a7b0b88fb6d99425154695c6c7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1969580
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#727978}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 63261b2bc636e66fc923fd0748681c160c6cd05f