[iOS] Add OverlaySupport specification to OverlayDispatchCallbacks
OverlayDispatchCallbacks were previously a typedef of
base::RepeatingCalback that was stored in
OverlayDispatchCallbackStorage, which used a templatized API in order
to ensure that the provided callbacks were only ever executed with a
specific type of OverlayResponse.  This added additional complexity
that could not be hidden because C++ class templates need to be defined
entirely within header files since they are not linked.

This CL instead defines an OverlayDispatchCallback class that allows
for the specification of a subset of allowed OverlayResponses, removing
the need for a templatized storage class.

Bug: none
Change-Id: Ia7bbc2dc90ebf8941e12126104b076a98a6feffa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1984177
Commit-Queue: Kurt Horimoto <kkhorimoto@chromium.org>
Reviewed-by: Mike Dougherty <michaeldo@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#728107}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: dbbe008f7bd15bde786ecf7e359ec025384a2c77