[iOS][web] Use common IsURLTrustworthy implementation
This CL uses the common implementation in /service to know if a URL
should be considered as secure or not, instead of having a custom
implementation for iOS.
It allows to manage the blob: scheme and reduces the code duplication.

This is doing a forward call instead of changing the call sites to
make sure that the test are passing. The call sites will be updated
later.

Previously trustworthy, now treated as not trustworthy:
- data scheme URLs (explicitly rejected by IsOriginPotentiallyTrustworthy's opaqueness checks + explicitly rejected by IsUrlPotentiallyTrustworthy).

Previously inner origin not extracted, now correct behavior:
- blob: URLs (because network::IsUrlPotentiallyTrustworthy calls url::Origin::Create)

Previously not trustworthy, now treated as trustworthy:
- file scheme (explicitly covered by network::IsOriginPotentiallyTrustworthy, but not by the old SchemeRegistry::secure_schemes)

No change in behavior:
- about scheme (covered by old SchemeRegistry::secure_schemes and explicitly handled by new network::IsUrlPotentiallyTrustworthy)
- https, wss scheme (covered both by old SchemeRegistry::secure_schemes and new GURL::SchemeIsCryptographic)
- quic (covered by old SchemeRegistry::secure_schemes and new url::GetSecureSchemes check from IsOriginPotentiallyTrustworthy)
- net::IsLocalhost called in the old and new code path

Bug: 939077
Change-Id: Ic2f12f9c64e845f78c26bc764f900e0e1108953c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1982540
Reviewed-by: Eugene But <eugenebut@chromium.org>
Reviewed-by: Łukasz Anforowicz <lukasza@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#728160}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fdad565fda83810ebbeb925c9523daeafe86d7d8