[iOS] Reenable flaky FontSizeJS test
I ran the test multiple times locally and wasn't able to reproduce the
flakiness. However, the error is the same as that observed if you
remove the viewport tag from |LoadHtml| and the same as the issue
described in that comment.

I tried adding in the -webkit-text-size-adjust flag in hopes that the
two methods together will prevent flakiness.

Fixed: 1038529
Change-Id: I742005ec22618566c6279d89fe9d1d7ca47c2867
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1983171
Reviewed-by: Eugene But <eugenebut@chromium.org>
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Cr-Original-Commit-Position: refs/heads/master@{#728208}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 84305e347bc7449bae0c31600b06b3e289e4cf04