[iOS] Create ios/chrome/browser/infobars/overlays/OWNERS
This CL adds kkhorimoto@chromium.org and sczs@chromium.org as OWNERS of
this directory.  kkhorimoto@ is the author of all the code added so far
to the directory, and sczs@ is the reviewer for the added code.

Bug: none
Change-Id: I4b3942b873ee4690e125788e34d244ac46af6cfe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1989016
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Commit-Queue: Sergio Collazos <sczs@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#728909}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4533c3b762fbc3d52ce40376b87e1748199d7a89