[iOS] Move Find Bar code from BVC to FindBarCoordinator
A lot of the code controlling Find In Page lived in the BVC. This CL
moves most of it out into a new FindBarCoordinator to control the
feature.

There are two places where code is left in BVC. The first is many
places where BVC uses FindTabHelper to check whether Find In Page is
enabled or not. The second is the implementation of |searchFindInPage|,
|findNextStringInPage|, and |findPreviousStringInPage| from
BrowserCommands. These are still in BVC because they are keyboard
shortcuts, and moving those around was more difficult.

Change-Id: Ie5dc882c98064b75069894d0aa246f9f8e6b0da3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1950540
Reviewed-by: Mark Cogan <marq@chromium.org>
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Cr-Original-Commit-Position: refs/heads/master@{#729311}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 33fa7c87a5339f672ddfad1d79a721550280abcf