[iOS] Add skeleton of Text Zoom UI
This adds a simple UI for text zoom (just the done button). The more
important contribution is that now, BVC will check which feature is
active, text zoom or find in page, and if something asks it to display
the other, it first dismisses the shown one.

Bug: 1028938
Change-Id: I576a096a635e0736f70e853f3c78e5459d7f4b78
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1946481
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#730153}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3b0649bfc925944488dbcdb59de022332083e134