Fix Cookies EG2 Test
Remove Assert in the ChromeEarlGrey Cookies method, which was there to
assert that the object returned from the javascript execution is iterable
object. The respondtoSelector method doesn't work properly in that case
and it returns always false even if the EdoObject is actually an Array.

Bug: 1019670, 1041000
Change-Id: I11f042d1ff2ffaeeabef1f39deffe6ed554d1e5f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1994101
Commit-Queue: Mohammad Refaat <mrefaat@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#730269}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9285d7a228e2b30966513bab5151976efb1031e7