Add JavaScript commands for iOS SSL committed interstitials
This CL adds a ScriptCommandSubscription to IOSSSLErrorTabHelper
to manage the command listener for SSL interstitials. This subscription
will listen for JS commands that come from the interstitial (for which JS
has been updated to properly pass messages for iOS) and forward commands
to the security interstitial subclass, which has a command handler method
implemented for the specific relevant cases. Most of the commands re-use
logic from non-committed interstitials, except for IOSSSLBlockingPage's
proceed command.

This CL also makes SessionCertificatePolicyCacheImpl::RegisterAllowedCertificate
public so that it can be used for allowlisting a cert when a user proceeds
on an SSL interstitial. This method is also updated to handle AllowCertForHost
so that only one method needs to be called for allowlisting.

Bug: 987407
Change-Id: I5b216cca66f2aebf0df8cec293b60ed0f862fae2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1937853
Commit-Queue: Livvie Lin <livvielin@chromium.org>
Reviewed-by: Kurt Horimoto <kkhorimoto@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Reviewed-by: Carlos IL <carlosil@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#730278}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d0af08bbd585ef593fd81fbb8a12bd433a764bf7