[iOS] Convert user-initiated banner dismissal response to dispatch
The purpose of this callback is to call
InfoBarDelegate::InfoBarDismissed(), which resets the state within the
delegate to record that the infobar banner was dismissed by tapping the
X in the legacy UI or swiping up to dismiss in the refresh UI.  This
needs to happen before the UI is actually dismissed, so that the
correct UMA metric can be recorded when the UI is dismissed.

Previously, this occurred in an OverlayRequest completion block, which is
executed when the OverlayRequest is completed (i.e. when the UI for the
request is guaranteed to never been shown again).  An OverlayRequest's
completion block is executed after observers are notified of the
completion of the UI's dismissal, meaning that the UMA metric is not
set properly by the time it's recorded.

This CL converts the banner completion response to a dispatch response
so that the UMA metric can be updated before the dismissal finishes.  It
also updates InfobarBannerInteractionHandler to handle this dispatched
response.  Now, when an infobar banner undergoes a user-initiated
dismissal, its handler first receives the BannerDismissedByUser()
callback before BannerVisibilityChanged().

Bug: 1030357
Change-Id: I3a762e61453030388e79f662ea32f0203dcdf77e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1991716
Commit-Queue: Kurt Horimoto <kkhorimoto@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#730312}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b4fa2cbaf2166941f701be9091861db1d7aede69