[iOS] Fix bugs causing zoom/find in page to show/hide at the wrong time
The major bug being fixed is that when navigating, |closeFindInPage| is
called twice, because both the FindTabHelper and the BVC listen for
the navigation event and stop finding. To fix this, neither place
listens for navigation any more, and FindBarMediator now handles that.

There were also places where find in page was being closed because
other things were happening and text zoom should be closed as well
(e.g. voice search starting).

Some of these places are simply hiding the UI (as opposed to stopping
FindTabHelper) because something is happening (device rotation/split
screen). For FindInPage, the FindTabHelper holds whether finding is in
progress, so the UI does come back. This doesn't exist yet for
TextZoom, but closing the UI too often is much better than leaving it
open when it shouldn't be.

Bug: 1028938, 918934
Change-Id: If916328883624fb591817bd79f5ae621ee57f308
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997305
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#731346}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fc8d0f3124c201b99da1a80d1baf3ffe49329c19