[ios] Fix broken autolayout constraints in tab grid bottom toolbar
There are broken autolayout constraints when starting the app.
The constraints break in iOS 13, though it is not visible to the
user. The issue is that UIToolbar needs to be initialized with
a non-zero rect in iOS 13.

Bug: 1029344
Change-Id: I7871bbe5c1b9d75c3970f336ccc90ad2337b04ac
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2001021
Reviewed-by: edchin <edchin@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Auto-Submit: edchin <edchin@chromium.org>
Commit-Queue: edchin <edchin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#731991}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1c993f68b13cf43fb0e99c9b9d29007f37d733e8