Reformat all gn files in /ios/chrome/browser/ui/main.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=edchin@chromium.org

Bug: 1041419
Change-Id: I9a1edf49df78480bf90f2eec8578f150a9bd21bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997658
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: edchin <edchin@chromium.org>
Commit-Queue: edchin <edchin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#732005}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: bb88b318ee0b63f5020748900fdf37494bbbbe29