[iOS] Support showing the passwords settings for infobar modals.
This CL updates PasswordInfobarModalInteractionHandler to take Browser
upon construction so that it can use the command dispatcher to open the
passwords settings when selected in the password infobar modal view.

Additionally, this CL adds default behavior to the show settings command
for use when there is not a base view controller to send with the
command.  OverlayRequestCoordinators' base view controllers are their
corresponding OverlayContainerViewController, which only should be used
to present things from within that OverlayPresenter's modality.  The
default behavior is to instead use the current BVC as the base controller
if one is not provided.

Bug: 1030357
Change-Id: I9a3d18cad6425aed47c005ce52ec7119b4a9059c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2001727
Commit-Queue: Kurt Horimoto <kkhorimoto@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#733784}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d798e5448fe77c8d42f3e36c05c18c2ff733ac6f