[iOS] Update WebContent correctly
This CL makes sure that the WebContent is taken into account when
requesting a page. This allows the WKWebView to know how to react to the
JavaScript event. In particular, this changes the navigator.platform
property. It adds the corresponding tests.

It also adds a new utility to find a string through all the frames of a
WebState.

Bug: 1031124
Change-Id: I2a794ca2bd8c7878d0676a2a9ec5d4269d68cb18
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1951010
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Ali Juma <ajuma@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#736335}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6783f497ff2fcfdf031b4d7b488862428c97e457