[ios] Remove FullscreenController as a KeyedService
This is an intermediate step to scoping FullscreenController to a Browser.
In order to accomplish this, the code needs to be able to switch between
associating it to both a BrowserState, which it current is, and a Browser.
FullscreenController will have two new static getter methods FromBrowser()
and FromBrowserState() to provide such abilities.

A new FullscreenControllerContainer class that will be associated to either
a BrowserState or Browser will own and manage the FullscreenControllerImpl
instance.

Since it is not used after this change, FullscreenControllerFactory is
removed.

Bug: 790886
Change-Id: I9a1e4e9832911d045b389b892538c00327c83127
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2006111
Commit-Queue: Chris Lu <thegreenfrog@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Reviewed-by: Kurt Horimoto <kkhorimoto@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#736571}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 09c205af532d8323b375a3fa0beb27af0de40784