[ios] Delete dead code (consumer)
This code is not build and not used. The README.txt says this is a
proof-of-concept, but as far as I can tell this has never been used
and is not maintained (since it is not build).

Bug: none
Change-Id: I42217423cd0e83189c8f59097ea876df036ca0f4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2027529
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#736888}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 328c515e1a8db22db973c220ba9151330e85a65e