Remove const feature manager in ChromePasswordManagerClient
Currently, the member password_feature_manager_ is marked const in the
ChromePasswordManagerClient but only ever accessed in a non-const
context.
This is possible since PasswordManagerClient::GetPasswordFeatureManager
casts the const away.

This CL removes the misleading const qualifier and the unused accessor
to the const method as well.
This clean-up is a precondition for the linked bug which needs non-const
access to the feature manager in ChromePasswordManagerClient to call
SetAccountStorageOptIn.

Bug: 1047202
Change-Id: Ib1383043e25843052815999e00cd163453442d64
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2030866
Commit-Queue: Friedrich [CET] <fhorschig@chromium.org>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#737291}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: bc4631656c0dc9e0b3e30a2eb324b4e25234a55e