[Payment Handler] Set request initiator for manifest fetches.
This patch passes the origin of the iframe that called the
PaymentRequest API to the fetcher of manifest files, which
uses it for sec-fetch-site headers. This patch is for service
worker based payment handlers on both Android and desktop.
This patch does not alter how iOS or native Android apps work.

Bug: 966507
Change-Id: I97b6ca08fd9e08d1762f95d7b78edf7d73450aee
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1636635
Commit-Queue: Rouslan Solomakhin <rouslan@chromium.org>
Reviewed-by: Łukasz Anforowicz <lukasza@chromium.org>
Reviewed-by: Danyao Wang <danyao@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#737484}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c71cca62949021fe70c170ce68a020fe407562d2