Opens DocumentPicker without providing a directory URL
- Opens the Document picker without providing a directory URL to avoid the disappearance of the Cancel button.
The cancel button seems to be hidden when the Downloads folder is empty for obscure reasons and it magically appears if you go to the "recents" section.
By not providing the directory URL, at the first use of the document picker, the user will be in the "recents" section.
During the following uses, the document picker will open where the user closed it for the last time.
This view controller is provided by Apple so we can't do much.

Bug: 1046329
Change-Id: I7bea3c69f0a9317f9cddda9a59ad8a91e7d91863
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2030964
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Ewann Pellé <ewannpv@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#738117}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 689362cb30820fd1a95b7ad9ad25ae81188ba9a5