[ios] Added manual password saving
In the case of manual password saving, we need to
skip the 'save password' infobar banner and let the user click
the key badge to manually save or update the password.

We also need to be able to remove the badge if user removes password.
I've kept the badge handling as is, tied with the banner and infobar,
even though we skip the banner.

Works only for info reboot ui.
Can be demo tested by patching 2012031 over this.

Bug: 1045431
Change-Id: I0b208467910311ab0c86c4b90da593bb1827f93c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2019267
Commit-Queue: David Jean <djean@chromium.org>
Reviewed-by: Kurt Horimoto <kkhorimoto@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Reviewed-by: Vadym Doroshenko <dvadym@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#738526}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: cb76c248b18a9554e041fd98ccd92f89e27f109e