Use EarlGrey wait condition instead of spinning the runloop.
The sign-in external URL tests were spinning the runloop while waiting
for the identities to be removed from the ChromeIdentityService. This
was consistently failing with EarlGrey2.

This CL uses an EarlGrey wait condition instead of spinning the runloop
while waiting for the identities to be removed.

Bug: 1031986

Change-Id: I0bcf6bda55caa506fe3a4d54ee607efc5c2e4d37
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2039551
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Mihai Sardarescu <msarda@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#738916}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7a4958d3316434f6a267bf66d1faa4af9ad2f430